Fixed acceleration structure scratch buffer alignment and modified materials texture index type
Added device extension VK_KHR_16bit_storage.
Material's texture type changed from uint64_t to uint16_t.

+++ include/material.h
@@ -11,7 +11,7 @@ namespace core {
 		~Material();
 
 		void cleanup();
-		void setup(uint64_t albedoMapIndex, uint64_t metallicMapIndex, uint64_t normalMapIndex);
+		void setup(uint16_t albedoMapIndex, uint16_t metallicMapIndex, uint16_t normalMapIndex);
 
 		Buffer& getBuffer() { return materialBuffer; }
 
@@ -25,9 +25,9 @@ namespace core {
 		glm::vec2 tilling;    // Vector 2D which scales all the mapped textures.
 		glm::vec2 offset;     // Vector 2D which offsets all the mapped textures.
 
-		uint64_t albedoMapIndex;
-		uint64_t metallicMapIndex;
-		uint64_t normalMapIndex;
+		uint16_t albedoMapIndex;
+		uint16_t metallicMapIndex;
+		uint16_t normalMapIndex;
 		Buffer materialBuffer; // On device buffer containing the material's data.
 
 		void createMaterialBuffer(Buffer& buffer);
